Prevent reprocessing known enumerations
Previously, for every property in every schema class, this was told to find every enumeration in the list of classes with a matching label. Instead, we can process the list once instead of N^3. A hashset also provides fast lookups of the data itself.

This alone dramatically improves performance of processing the schema data.